Improve objprint.fn_8003B0D0 98.78->100.00: int maxAngle param (drops extsh-before-multiply u32->double conv; callers pass int 0x19) + drop the (int) middle cast on the reassignment so (s16)(float*maxAngle) coalesces into maxAngle's own param home r30 (the (int) forced an r3 temp)

+++ src/main/objprint.c
@@ -1944,7 +1944,7 @@ void fn_8003A230(int obj, void* state, f32 val)
 extern int getAngle(float y, float x);
 extern f32 gObjPrintDegToAngle;
 
-void fn_8003B0D0(int obj, int target, int state, s16 maxAngle)
+void fn_8003B0D0(int obj, int target, int state, int maxAngle)
 {
     s16* found;
 
@@ -1956,7 +1956,7 @@ void fn_8003B0D0(int obj, int target, int state, s16 maxAngle)
                                                         ((GameObject*)obj)->anim.localPosZ -
                                                             *(f32*)((char*)target + 0x14)) -
                                           ((GameObject*)obj)->anim.rotX);
-        maxAngle = (s16)(int)(gObjPrintDegToAngle * maxAngle);
+        maxAngle = (s16)(gObjPrintDegToAngle * maxAngle);
         if (*(s16*)((char*)state + 0x14) > maxAngle)
         {
             *(s16*)((char*)state + 0x14) = maxAngle;
